refactor: Implement init function in on_audio module for explicit side effects
